pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "forge-cli"
version = "0.1.0"
description = "Self-managing AI agent CLI (forge)"
requires-python = ">=3.11"
authors = [{name = "ai"}]
dependencies = [
  "click>=8.1",
  "rich>=13.7",
  "pyyaml>=6.0",
  "anthropic>=0.40",
  "aiosqlite>=0.20",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
  "black>=24.8",
  "mypy>=1.11",
  "types-PyYAML",
]

[project.scripts]
forge = "forge.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = {find = {where = ["src"]}}

[tool.ruff]
line-length = 100
select = ["E","F","I","UP","B"]
ignore = ["E501"]

[tool.black]
line-length = 100

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = false
warn_unused_ignores = true
ignore_missing_imports = true
